+++ app/Http/Controllers/Api/EconomicoController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Http\Requests\StoreEconomicoRequest;
 use App\Http\Requests\UpdateEconomicoRequest;
 use App\Http\Resources\EconomicoResource;
@@ -11,16 +12,19 @@ use App\Models\Economia;
 
 class EconomicoController extends Controller
 {
+    use AuthorizesRequests;
     //Vista todos los datos
     public function index()
-    {
+    {   
+        $this->authorize('Ver registros');
         $economico = Economia::all();
         return EconomicoResource::collection($economico);
     }
 
     //Crear un nuevo registro
     public function store(StoreEconomicoRequest $request)
-    {
+    {   
+        $this->authorize('CrearEliminar');
         $economia = Economia::create($request->validated());
         return response()->json($economia, 201);
        return new EconomicoResource($economia);
@@ -31,7 +35,8 @@ class EconomicoController extends Controller
      * Display the specified resource.
      */
     public function show(Economia $economico)
-    {
+    {   
+        $this->authorize('Ver registros');
         return new EconomicoResource($economico);
     }
 
@@ -39,7 +44,8 @@ class EconomicoController extends Controller
      * Update the specified resource in storage.
      */
     public function update(UpdateEconomicoRequest $request, Economia $economico)
-    {
+    {   
+        $this->authorize('Modificar registros');
         $economico->update($request->validated());
         return new EconomicoResource($economico);
     }
@@ -48,7 +54,8 @@ class EconomicoController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(string $id)
-    {
+    {   
+        $this->authorize('CrearEliminar');
         $economia = Economia::findOrFail($id);
         $economia->delete();
         return response()->json(['message' => 'Registro eliminado correctamente']);
